finished extrusion example and removed images from nav drawer
