+++ MapboxAndroidDemo/src/main/java/com/mapbox/mapboxandroiddemo/examples/extrusions/BasicExtrusionActivity.java
@@ -1,4 +1,4 @@
-package com.mapbox.mapboxandroiddemo.examples.styles;
+package com.mapbox.mapboxandroiddemo.examples.extrusions;
 
 import android.graphics.Color;
 import android.os.Bundle;
