--- MapboxAndroidDemo/src/main/res/drawable/ic_dds_24dp.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-        android:width="24dp"
-        android:height="24dp"
-        android:viewportWidth="24.0"
-        android:viewportHeight="24.0">
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M3,16h5v-2L3,14v2zM9.5,16h5v-2h-5v2zM16,16h5v-2h-5v2zM3,20h2v-2L3,18v2zM7,20h2v-2L7,18v2zM11,20h2v-2h-2v2zM15,20h2v-2h-2v2zM19,20h2v-2h-2v2zM3,12h8v-2L3,10v2zM13,12h8v-2h-8v2zM3,4v4h18L21,4L3,4z"/>
-</vector>
