+++ MapboxAndroidDemo/src/main/res/layout/activity_basic_extrusion.xml
@@ -1,22 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:mapbox="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.mapbox.mapboxandroiddemo.examples.styles.BasicExtrusionActivity">
+    tools:context="com.mapbox.mapboxandroiddemo.examples.extrusions.BasicExtrusionActivity">
 
     <com.mapbox.mapboxsdk.maps.MapView
         android:id="@+id/mapView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        mapbox:mapbox_cameraBearing="82.104001"
-        mapbox:mapbox_cameraTargetLat="52.516318"
-        mapbox:mapbox_cameraTargetLng="13.378105"
-        mapbox:mapbox_cameraTilt="60.00"
-        mapbox:mapbox_cameraZoom="16.922091"
+        mapbox:mapbox_cameraBearing="300"
+        mapbox:mapbox_cameraTargetLat="41.87827"
+        mapbox:mapbox_cameraTargetLng="-87.62877"
+        mapbox:mapbox_cameraTilt="60"
+        mapbox:mapbox_cameraZoom="15"
         mapbox:mapbox_styleUrl="@string/mapbox_style_mapbox_streets"/>
 
 </RelativeLayout>
