+++ MapboxAndroidDemo/src/main/res/layout/activity_population_density_extrusion.xml
@@ -2,16 +2,14 @@
 <android.support.constraint.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:fab="http://schemas.android.com/apk/res-auto"
     xmlns:mapbox="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.mapbox.mapboxandroiddemo.examples.styles.PopDensityExtrusionActivity">
+    tools:context="com.mapbox.mapboxandroiddemo.examples.extrusions.PopulationDensityExtrusionActivity">
 
     <com.mapbox.mapboxsdk.maps.MapView
         android:id="@+id/mapView"
-        mapbox_uiTiltGestures="false"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -20,70 +18,27 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0"
-        mapbox:mapbox_cameraTargetLat="37.775569"
-        mapbox:mapbox_cameraTargetLng="-122.410973"
+        mapbox:mapbox_cameraTargetLat="37.784282779035216"
+        mapbox:mapbox_cameraTargetLng="-122.4232292175293"
         mapbox:mapbox_cameraZoom="12"
         mapbox:mapbox_cameraZoomMax="16"
         mapbox:mapbox_cameraZoomMin="12"
+        mapbox:mapbox_cameraTilt="45"
         mapbox:mapbox_styleUrl="@string/mapbox_style_dark">
     </com.mapbox.mapboxsdk.maps.MapView>
-
-    <com.getbase.floatingactionbutton.FloatingActionsMenu
-        android:id="@+id/multiple_map_actions"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="end|bottom"
-        android:layout_marginBottom="16dp"
-        android:layout_marginEnd="16dp"
-        android:layout_marginRight="16dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        fab:fab_addButtonColorNormal="@color/mapboxRed"
-        fab:fab_addButtonColorPressed="@color/mapboxWhite"
-        fab:fab_addButtonPlusIconColor="@color/mapboxWhite"
-        fab:fab_labelStyle="@style/menu_labels_style">
-
-        <com.getbase.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab_visualization_toggle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            fab:fab_colorNormal="@color/mapboxBlue"
-            fab:fab_colorPressed="@color/mapboxWhite"
-            fab:fab_size="mini"
-            fab:fab_title="@string/toggle_map_tilt_map_toggle_fab_text" />
-
-        <com.getbase.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab_road_toggle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            fab:fab_colorNormal="@color/mapboxGreen"
-            fab:fab_colorPressed="@color/mapboxWhite"
-            fab:fab_size="mini"
-            fab:fab_title="@string/toggle_map_road_toggle_fab_text" />
-
-        <com.getbase.floatingactionbutton.FloatingActionButton
-            android:id="@+id/fab_label_toggle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            fab:fab_colorNormal="@color/mapboxOrange"
-            fab:fab_colorPressed="@color/mapboxWhite"
-            fab:fab_size="mini"
-            fab:fab_title="@string/toggle_map_label_toggle_fab_text" />
-
-    </com.getbase.floatingactionbutton.FloatingActionsMenu>
 
     <android.support.constraint.Guideline
         android:id="@+id/guideline4"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        app:layout_constraintGuide_begin="20dp" />
+        app:layout_constraintGuide_begin="20dp"/>
 
     <android.support.constraint.Guideline
         android:id="@+id/guideline5"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        app:layout_constraintGuide_begin="20dp" />
+        app:layout_constraintGuide_begin="20dp"/>
 
 </android.support.constraint.ConstraintLayout>
